feat: add room and roomtype endpoints

+++ backend/src/main/java/com/rayan/cinemaapi/service/RoomService.java
@@ -1,6 +1,7 @@
 package com.rayan.cinemaapi.service;
 
 import com.rayan.cinemaapi.entity.Room;
+import com.rayan.cinemaapi.entity.RoomType;
 import com.rayan.cinemaapi.exception.EntityNotFoundException;
 import com.rayan.cinemaapi.repository.RoomRepository;
 import org.springframework.stereotype.Service;
@@ -12,9 +13,11 @@ import java.util.List;
 public class RoomService {
 
     private final RoomRepository roomRepository;
+    private final RoomTypeService roomTypeService;
 
-    public RoomService(RoomRepository roomRepository) {
+    public RoomService(RoomRepository roomRepository, RoomTypeService roomTypeService) {
         this.roomRepository = roomRepository;
+        this.roomTypeService = roomTypeService;
     }
 
     public List<Room> getRooms(){
@@ -28,10 +31,20 @@ public class RoomService {
     }
 
     public Room createRoom(Room room) {
+        RoomType roomType = roomTypeService.getRoomType(room.getRoomType().getId());
+
+        // Only the id of the roomtype gets supplied in a request
+        // The following line gets used to fill in the other attributes of the roomtype
+        room.setRoomType(roomType);
+
         return roomRepository.save(room);
     }
 
     public void deleteRoom(Long id) {
+        if (!roomRepository.existsById(id)) {
+            throw new EntityNotFoundException("Room", id);
+        }
+
         roomRepository.deleteById(id);
     }
 
@@ -39,7 +52,10 @@ public class RoomService {
     public Room updateRoom(Room room) {
         Room existingRoom = getRoom(room.getId());
 
-        existingRoom.setRoomType(room.getRoomType());
+        RoomType roomType = roomTypeService.getRoomType(room.getRoomType().getId());
+
+        existingRoom.setName(room.getName());
+        existingRoom.setRoomType(roomType);
 
         return existingRoom;
     }
